split component into data container and view
