+++ assets/prototype/domain/eventEditor/tickets/ticketPriceCalculator/TicketPriceCalculatorButton.js
@@ -1,19 +1,11 @@
-import { useContext } from 'react';
 import { Button } from '@blueprintjs/core/lib/esm';
 import TicketPriceCalculatorModal from './TicketPriceCalculatorModal';
-import { TicketContext } from '../../context/TicketProvider';
 
-const TicketPriceCalculatorButton = ({ ticket }) => {
-	const { getIsOpen, onClose, setIsOpen } = useContext(TicketContext);
-	const modalId = 'TicketPriceCalculator';
-	const isOpen = getIsOpen(modalId);
-	const onOpen = () => setIsOpen(modalId);
-	const handleClose = () => onClose(modalId);
-
+const TicketPriceCalculatorButton = ({ ticket, isOpen, onOpen, onClose }) => {
 	return (
 		<>
-			<Button icon={'calculator'} onClick={onOpen} minimal />
-			<TicketPriceCalculatorModal ticket={ticket} handleClose={handleClose} isOpen={isOpen} />
+			<Button icon={'calculator'} onClick={() => onOpen()} minimal />
+			<TicketPriceCalculatorModal ticket={ticket} handleClose={() => onClose()} isOpen={isOpen} />
 		</>
 	);
 };
